Dropped the return value and statement-by-default
Now mixin(Loop!"...") compiles to a statement containing the generated code and mixin(Loope!"...") compiles to an expression ready to be assigned to a variable or passed to a function.
All type checking and type deduction is left to D, so expect all kinds of weird errors untill this is resolved (if ever).